scripts/seo/requirements.txt
# SEO Automation Suite - Python Dependencies
# Install: pip install -r scripts/seo/requirements.txt

# Environment & config
python-dotenv>=1.0.0

# Template engine (page generator)
jinja2>=3.1.0

# HTTP requests (freshness monitor, link checker)
requests>=2.31.0

# Data analysis (dashboard, reports)
pandas>=2.0.0

# Google APIs (GSC + GA4)
google-api-python-client>=2.114.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-analytics-data>=0.20.0

# LLM APIs (visibility tracker) - optional
openai>=1.0.0
anthropic>=0.20.0

# Video pipeline (Slack + Google Sheets)
slack_sdk>=3.27.0
gspread>=6.0.0

# Thumbnail generation
Pillow>=10.0.0
